requirements.txt
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.12.2
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
confection==0.0.4
cymem==2.0.7
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.5.0/en_core_web_trf-3.5.0-py3-none-any.whl
exceptiongroup==1.1.1
filelock==3.11.0
h11==0.14.0
huggingface-hub==0.13.4
idna==3.4
Jinja2==3.1.2
langcodes==3.3.0
Levenshtein==0.20.9
MarkupSafe==2.1.2
mpmath==1.3.0
murmurhash==1.0.9
networkx==3.1
numpy==1.24.2
outcome==1.2.0
packaging==23.1
pathy==0.10.1
preshed==3.0.8
pydantic==1.10.7
PySocks==1.7.1
python-dotenv==1.0.0
python-Levenshtein==0.20.9
PyYAML==6.0
rapidfuzz==2.15.1
regex==2023.3.23
requests==2.28.2
requests-file==1.5.1
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
spacy==3.5.2
spacy-alignments==0.9.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
spacy-transformers==1.2.2
srsly==2.4.6
sympy==1.11.1
thefuzz==0.19.0
thinc==8.1.9
thinc-apple-ops==0.1.3
tldextract==3.4.0
tokenizers==0.13.3
torch==2.0.0
tqdm==4.65.0
transformers==4.26.1
trio==0.22.0
trio-websocket==0.10.2
typer==0.7.0
typing_extensions==4.5.0
urllib3==1.26.15
wasabi==1.1.1
wsproto==1.2.0
